Add files via upload
The two-sample t-test was used here to compare the two groups of undamaged vehicles and damaged vehicles. The means of each group and the difference between them are in the third and fourth tables in figure 39. In figure 39, the t-test based on a pooled estimate of the variance has a p-value of less than 0.0001 and for the Satterthwaite version, the p-value is also less than 0.0001. Both p-values are less than 0.05, so both suggest rejecting the null hypothesis for the third question. The F-test is used to determine if the variances can be equal. The p-value for this test is also less than 0.0001, then the variances are not equal, and it appears that the variance for a damaged vehicle is larger than that for the undamaged vehicle. It can conclude that the annual premium is significantly different between undamaged vehicles and damaged vehicles.